Application/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
bcrypt
pyjwt
sendgrid
python-dotenv